# internal/api/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Enduro API
  version: ""
host: localhost:9000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /pipeline:
    get:
      tags:
      - pipeline
      summary: list pipeline
      description: List all known pipelines
      operationId: pipeline#list
      parameters:
      - name: name
        in: query
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: array
            items:
              $ref: '#/definitions/EnduroStoredPipelineResponse'
      schemes:
      - http
  /pipeline/{id}:
    get:
      tags:
      - pipeline
      summary: show pipeline
      description: Show pipeline by ID
      operationId: pipeline#show
      parameters:
      - name: id
        in: path
        description: Identifier of pipeline to show
        required: true
        type: string
        format: uuid
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/PipelineShowResponseBody'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/PipelineShowNotFoundResponseBody'
            required:
            - message
            - id
      schemes:
      - http
  /pipeline/{id}/processing:
    get:
      tags:
      - pipeline
      summary: processing pipeline
      description: List all processing configurations of a pipeline given its ID
      operationId: pipeline#processing
      parameters:
      - name: id
        in: path
        description: Identifier of pipeline
        required: true
        type: string
        format: uuid
      responses:
        "200":
          description: OK response.
          schema:
            type: array
            items:
              type: string
              example: Voluptates et molestiae eum et occaecati.
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/PipelineProcessingNotFoundResponseBody'
            required:
            - message
            - id
      schemes:
      - http
...
definitions:
...
  EnduroStoredPipelineResponse:
    title: 'Mediatype identifier: application/vnd.enduro.stored-pipeline; view=default'
    type: object
    properties:
      capacity:
        type: integer
        description: Maximum concurrent transfers
        example: 2530439535807859749
        format: int64
      current:
        type: integer
        description: Current transfers
        example: 5707507152436743067
        format: int64
      id:
        type: string
        description: Identifier of the pipeline
        example: 73C3FFFF-5E99-6BB1-43A8-61CBFDE01330
        format: uuid
      name:
        type: string
        description: Name of the pipeline
        example: Et magni consequuntur.
    description: StoredPipeline describes a pipeline retrieved by this service. (default
      view)
    example:
      capacity: 1857366314510706938
      current: 6412453802971423915
      id: 9CEC8109-7649-F74C-CDA6-D8C4434F6E6F
      name: Tenetur quibusdam rerum sit labore.
    required:
    - name
  PipelineProcessingNotFoundResponseBody:
    title: PipelineProcessingNotFoundResponseBody
    type: object
    properties:
      id:
        type: string
        description: Identifier of missing pipeline
        example: Fugiat a officiis ipsum.
      message:
        type: string
        description: Message of error
        example: Et et eligendi voluptatem.
    description: Pipeline not found
    example:
      id: Qui rerum asperiores qui.
      message: Necessitatibus voluptas mollitia dolore.
    required:
    - message
    - id
  PipelineShowNotFoundResponseBody:
    title: PipelineShowNotFoundResponseBody
    type: object
    properties:
      id:
        type: string
        description: Identifier of missing pipeline
        example: Praesentium sunt ut.
      message:
        type: string
        description: Message of error
        example: Facilis magnam a veritatis vel.
    description: Pipeline not found
    example:
      id: Voluptates nulla autem et perspiciatis libero voluptas.
      message: Illum illum tenetur.
    required:
    - message
    - id
  PipelineShowResponseBody:
    title: 'Mediatype identifier: application/vnd.enduro.stored-pipeline; view=default'
    type: object
    properties:
      capacity:
        type: integer
        description: Maximum concurrent transfers
        example: 1270080036653766538
        format: int64
      current:
        type: integer
        description: Current transfers
        example: 3460011645440185538
        format: int64
      id:
        type: string
        description: Identifier of the pipeline
        example: FDCA5323-C4C9-FB10-0E70-1125891ABDEE
        format: uuid
      name:
        type: string
        description: Name of the pipeline
        example: Ut tenetur dolorum error aut quibusdam.
    description: ShowResponseBody result type (default view)
    example:
      capacity: 4816212125257195638
      current: 5231338353180703567
      id: B79E857E-CE20-97AA-CE10-8106502AEA86
      name: Non unde ut aspernatur.
    required:
    - name